Move common stuff of c8y/tedge log manager to separate lib
