+++ crates/common/log_manager/src/error.rs
